Create the expansions

this function is used in each step of the Weng optimization and for the initialization of the XiaoXu optimization

DcFromDp(
  Dp,
  s,
  ell,
  permlist = rep(list(rep(list(rep(0:(ell - 1), each = nrow(Dp)/(s * ell))), s)),
    ncol(Dp))
)

Arguments

Dp

an OA (ideally maximin optimized GMA OA)

s

the number of levels of each column in Dp

ell

ell*s is the target number of levels of the outgoing Dc

permlist

a list of m lists of length s permutations of the elements of the replacement vector

Value

a matrix of the same size as Dp