Function to create maximin distance level expanded arrays

Maximin distance level expansion similar to Xiao and Xu is implemented, using an optimization algorithm that is less demanding than the TA algorithm of Xiao and Xu

MDLEs(
  oa,
  ell,
  noptim.rounds = 1,
  optimize = TRUE,
  noptim.oa = 1,
  dmethod = "manhattan",
  p = 50,
  storeperms = FALSE
)

Arguments

oa
matrix or data.frame that contains an ingoing symmetric OA. Levels must be denoted as 0 to s-1 or as 1 to s.

ell
the multiplier for each number of levels

noptim.rounds
the number of optimization rounds; optimization may take very long, therefore the default is 1, although more rounds are beneficial.

optimize
logical: if FALSE, suppress optimization of expansion levels

noptim.oa
integer: number of optimization rounds applied to initial oa itself before starting expansion

dmethod
distance method for phi_p, "manhattan" (default) or "euclidean"

p
p for phi_p (the larger, the closer to maximin distance)

storeperms
logical: should candidate permutations be stored? This can blow up the size of the output object enormously.

Value

A matrix of class MDLE with attributes

phi_p

the phi_p value that was achieved

type

MDLE

optimized

logical: same as the input parameter

call

the call that produced the matrix

permpick

matrix that lists the id numbers of the permutations used

perms2pickfrom

optional element, when optimization was conducted: the overall permutation list to which the numbers in permlist refer

Details

The ingoing oa is possibly optimized for space-filling, using function phi_optimize with noptim.oa optimization rounds. The expansions themselves are again optimized for improving phi_p, using an algorithm which is a variant of Weng (2014), instead of the more powerful but also much more demanding algorithm proposed by Xiao and Xu.

References

For full detail, see SOAs-package.

Weng (2014)
Xiao and Xu (2018)

Author

Ulrike Groemping

Examples

dim(MDLEs(DoE.base::L16.4.5, 2, noptim.rounds = 1))

#> Optimization round 1 of 1 started

#> [1] 16  5