Auxiliary function for optimized creation of OSOAs (function OSOAs) using the Li et al. algorithm for arbitrary initial OA

The optimization is done with function NeighbourcalcUniversal.R

OSOAarbitrary(oa, el = 3, m = NULL, permlist = NULL, random = TRUE)

Arguments

oa

matrix (preferred) or data.frame

el

must be 2 or 3. el=3: Li et al; el=2: Zhou and Tang

m

desired number of columns. Defaults to the maximum possible.

permlist

a list of length m of lists of length 1 of permutations of the levels

random

a logical (a random draw of permutations is used if permlist is NULL)

Value

an orthogonal array

References

Li et al. Zhou and Tang