function to create SOAs of strength 2+ from regular s-level designs

creates an array in s^k runs with columns in s^2 levels for prime or prime power s

SOAs2plus_regular(
  s,
  k,
  m = NULL,
  noptim.rounds = 1,
  noptim.repeats = 1,
  optimize = TRUE,
  dmethod = "manhattan",
  p = 50
)

Arguments

s
prime or prime power

k
array will have n=s^k runs; for s=2, k>=4 is needed, for s>2, k>=3 is sufficient

m
optional integer: number of columns requested; if NULL, the maximum possible number of columns is created, which is (s^k-1)/(s-1) - ((s-1)^k-1)/(s-2) for s>2 and s^k-s^k1 - s^(k-k1) + 2, with k1=floor(k/2), for s=2; specifying a smaller m is beneficial not only for run time but also for possibly achieving a column-orthogonal array (see Details section)

noptim.rounds
the number of optimization rounds for each independent restart

noptim.repeats
the number of independent restarts of optimizations with noptim.rounds rounds each

optimize
logical: should optimization be applied? default TRUE

dmethod
method for the distance in phi_p, "manhattan" (default) or "euclidean"

p
p for phi_p (the larger, the closer to maximin distance)

Value

matrix of class SOA with the attributes that are listed below. All attributes can be accessed using function attributes, or individual attributes can be accessed using function attr. These are the attributes:

type

the type of array (SOA or OSOA)

strength

character string that gives the strength

phi_p

the phi_p value (smaller=better)

optimized

logical indicating whether optimization was applied

permpick

matrix that lists the id numbers of the permutations used

perms2pickfrom

optional element, when optimization was conducted: the overall permutation list to which the numbers in permlist refer

call

the call that created the object

Details

The construction is by He, Cheng and Tang (2018), Prop.1 (C2) / Theorem 2 for s=2 and Theorem 4 for s>2.
B is chosen as an OA of strength 2, if possible, which yields orthogonal columns according to Zhou and Tang (2019). This is implemented using a matching algorithm for bipartite graphs from package igraph; the smaller m, the more likely that orthogonality can be achieved. However, strength 2+ SOAs are not usually advisable for m small enough that a strength 3 OA exists.
Optimization according to Weng has been added (separate level permutations in columns of A and B, noptim.rounds times). Limited tests suggest that a single round (noptim.rounds=1) often does a very good job (e.g. for s=2 and k=4), and further rounds do not yield too much improvement; there are also cases (e.g. s=5 with k=3), for which the unoptimized array has a better phi_p than what can be achieved by most optimization attempts from a random start.

Note

Strength 2+ SOAs can accommodate a large number of factors with reasonable stratified balance behavior. Note that their use is not usually advisable for m small enough that a strength 3 OA with s^2 level factors exists.

References

For full detail, see SOAs-package.

He, Cheng and Tang (2018)
Weng (2014)
Zhou and Tang (2019)

Author

Ulrike Groemping

Examples

## unoptimized SOA with 20 9-level columns in 81 runs
## (up to 25 columns are possible)
plan <- SOAs2plus_regular(3, 4, 20, optimize=FALSE)
## many column pairs have only 27 level pairs covered
count_npairs(plan)

#> $paircounts
#>   [1] 27 27 27 27 27 27 27 81 27 27 27 27 27 27 27 81 81 81 81 27 27 27 27 27 27
#>  [26] 81 81 81 27 27 81 81 81 27 27 27 27 27 27 27 27 27 27 81 81 27 27 81 81 81
#>  [51] 81 81 81 81 27 27 81 81 81 27 81 81 81 27 27 81 27 27 81 27 27 81 81 27 27
#>  [76] 81 81 81 27 27 81 81 81 27 81 81 81 27 81 27 81 81 81 81 27 27 27 81 27 27
#> [101] 81 27 27 27 27 81 81 81 27 27 27 81 81 27 27 27 27 81 81 81 81 81 81 27 81
#> [126] 81 27 81 81 27 81 81 27 27 81 27 81 81 27 27 81 81 27 81 81 81 81 81 81 27
#> [151] 27 81 27 81 27 27 27 27 27 27 27 81 27 27 27 81 81 81 27 27 27 27 81 81 27
#> [176] 27 81 27 27 81 27 81 81 81 27 27 27 27 27 81
#> 
#> $columnpaircounts
#>  [1]  783  837  999  945  999  945  945 1053 1161 1053 1107  837  999  999  945
#> [16] 1107  945  945 1053 1107
#> 

## an OA would exist for 10 9-level factors (DoE.base::L81.9.10)
## it would cover all pairs

## unoptimized OSOA with 8 16-level columns in 64 runs
## (maximum possible number of columns)
plan64 <- SOAs2plus_regular(4, 3, optimize=FALSE)
ocheck(plan64)   ## the array has indeed orthogonal columns

#> [1] TRUE