Implementation of the Xiao Xu TA algorithm (experimental, for comparison with MDLEs only)

Implementation of the Xiao Xu TA algorithm (experimental, for comparison with MDLEs only)

XiaoXuMDLE(
  oa,
  ell,
  noptim.oa = 1,
  nseq = 2000,
  nrounds = 50,
  nsteps = 3000,
  dmethod = "manhattan",
  p = 50
)

createF(Dc, Dp, s, ell, nseq = 2000)

optimize(
  Dc,
  s,
  ell,
  Fhat,
  nrounds = 50,
  nsteps = 3000,
  dmethod = "manhattan",
  p = 50
)

Arguments

oa

matrix or data.frame that contains an ingoing symmetric OA. Levels must be denoted as 0 to s-1 or as 1 to s.

ell

the multiplier for each number of levels

noptim.oa

integer: number of optimization rounds applied to initial oa itself before starting expansion

nseq

tuning parameters for TA algorithm

nrounds

tuning parameters for TA algorithm

nsteps

tuning parameters for TA algorithm

dmethod

distance method for phi_p, "manhattan" (default) or "euclidean"

p

p for phi_p (the larger, the closer to maximin distance)

Dc

matrix

Dp

matrix

s

original number of levels

Fhat

distribution function (created with createF)

Value

XiaoXuMDLE returns a matrix with attribute phi_p. createF returns a distribution function. optimize returns a matrix with attribute phi_p.

Details

The ingoing oa is optimized by function phi_optimize, using noptim.rounds=noptim.oa; this yields the matrix Dp for use in the internal functions DcFromDp and createF.
Function XiaoXuMDLE returns the value that is produced by applying the internal function optimize to the resulting Dc and F.

References

For full detail, see SOAs-package.

Xiao and Xu (2018)

Examples

## create 8-level columns from 4-level columns
XiaoXuMDLE(DoE.base::L16.4.5, 2, nrounds = 5, nsteps=50)
#>    A B C D E
#> 1  3 1 1 0 1
#> 2  2 3 4 3 7
#> 3  3 6 6 5 4
#> 4  2 4 3 7 2
#> 5  0 0 4 4 2
#> 6  1 2 0 7 5
#> 7  1 7 2 1 6
#> 8  0 5 6 2 1
#> 9  4 0 7 6 6
#> 10 5 2 2 5 0
#> 11 5 6 1 2 3
#> 12 4 5 5 0 5
#> 13 7 1 3 3 4
#> 14 6 3 7 1 3
#> 15 7 7 5 6 0
#> 16 6 4 0 4 7
#> attr(,"origin")
#> [1] "Kuhfeld collection"
#> attr(,"class")
#> [1] "matrix" "array" 
#> attr(,"phi_p")
#> [1] 0.1136378