Functions for Galois field calculations

Functions for Galois field calculations

int2poly(x, gf)

poly2int(poly, gf)

gf_sum(x, y, gf)

gf_prod(x, y, gf)

gf_sum_list(ll, gf, checks = TRUE)

gf_matmult(M1, M2, gf, checks = TRUE)

Arguments

x

a vector of e elements in 0:(q-1)

gf

Galois field object with characteristic q

poly

takes an e x n matrix of polygon rows or a length n vector for a single polygon (e=1)

y

a second vector of e elements in 0:(q-1)

ll

a list of integer vectors in 0:(q-1)

checks

logical: should checks be performed

M1

matrix 1 with elements in 0:(q-1)

M2

matrix 2 with elements in 0:(q-1)

Value

int2poly returns an e x n matrix or a length n vector (if x is scalar) poly2int returns an e element vector of integers in 0:(q-1)gf_sum returns the sum (e elements in 0:(q-1)) gf_prod returns the product (e elements in 0:(q-1)) gf_sum_list returns the sum (an integer vector in 0:(q-1)) gf_matmult returns the matrix product with elements from 0:(q-1)