function to create a strength 3 OSOA with 8-level columns from a Hadamard matrix

A Hadamard matrix in m runs is used for creating an OSOA in n=2m runs for at most m-2 columns.

OSOApb(m = NULL, n = NULL, el = 3)

Arguments

m
the number of columns to be created; if n is also given, m must be compatible with it

n
the number of runs to be created (must be a multiple of 8); if m is also given, n must be compatible with it

el
exponent for 2, can be 2 or 3: the OSOA will have columns with 2^el (4 or 8) levels

Value

an OSOA of strength 3- or 3 (matrix)

Details

At least one of m or n must be provided. For el=2, Zhou and Tang (2019) strength 3- designs are created, for el=3 strength 3 designs by Li, Liu and Yang (2021).

Note

Replaced by OSOAs_hadamard; eventually remove

References

Li, Liu and Yang (2021)

Author

Ulrike Groemping

Examples

dim(OSOApb(9))  ## 9 8-level factors in 24 runs

#> [1] 24  9

dim(OSOApb(n=16)) ## 6 8-level factors in 16 runs

#> Warning: Plackett-Burman designs in 8 runs coincide with regular fractional factorials. 
#>           For screening more than four factors, you may want to consider increasing the number of runs to 12. 
#>           Make sure to take the alias structure into account for interpretation!

#> [1] 16  6

dim(OSOApb(m=35)) ## 35 8-level factors in 80 runs

#> [1] 80 35