Implementation of the Xiao Xu TA algorithm (experimental, for comparison with MDLEs only)

Implementation of the Xiao Xu TA algorithm (experimental, for comparison with MDLEs only)

XiaoXuMDLE(
  oa,
  ell,
  noptim.oa = 1,
  nseq = 2000,
  nrounds = 50,
  nsteps = 3000,
  dmethod = "manhattan",
  p = 50
)

createF(Dc, Dp, s, ell, nseq = 2000)

optimize(
  Dc,
  s,
  ell,
  F,
  nrounds = 50,
  nsteps = 3000,
  dmethod = "manhattan",
  p = 50
)

Arguments

oa
matrix or data.frame that contains an ingoing symmetric OA. Levels must be denoted as 0 to s-1 or as 1 to s.

ell
the multiplier for each number of levels

noptim.oa
integer: number of optimization rounds applied to initial oa itself before starting expansion

nseq
tuning parameters for TA algorithm

nrounds
tuning parameters for TA algorithm

nsteps
tuning parameters for TA algorithm

dmethod
distance method for phi_p, "manhattan" (default) or "euclidean"

p
p for phi_p (the larger, the closer to maximin distance)

Dc
matrix

Dp
matrix

s
original number of levels

F
distribution function (created with createF)

Value

XiaoXuMDLE returns a matrix with attribute phi_p.

createF returns a distribution function.

optimize returns a matrix with attribute phi_p.

Details

The algorithm deviates from Xiao and Xu (2018) by optimizing the ingoing OA for phi_p instead of for the GWLP. Function XiaoXuMDLE returns the value that is produced by optimize using the distribution function calculated by createF.

References

For full detail, see SOAs-package.

Xiao and Xu (2018)

Examples

## create 8-level columns from 4-level columns
XiaoXuMDLE(DoE.base::L16.4.5, 2, nrounds = 5, nsteps=50)

#>    A B C D E
#> 1  0 0 2 4 1
#> 2  1 7 5 0 5
#> 3  0 3 7 7 6
#> 4  1 5 1 2 3
#> 5  5 1 5 6 3
#> 6  4 7 3 3 6
#> 7  4 2 0 5 4
#> 8  5 4 7 1 0
#> 9  7 0 6 2 4
#> 10 6 6 1 6 1
#> 11 7 3 3 0 2
#> 12 6 4 4 5 7
#> 13 2 1 0 1 7
#> 14 2 6 6 4 2
#> 15 3 2 4 3 0
#> 16 3 5 2 7 5
#> attr(,"origin")
#> [1] "Kuhfeld collection"
#> attr(,"class")
#> [1] "matrix" "array" 
#> attr(,"phi_p")
#> [1] 0.1158475