Utility function for inspecting SOAs obtainable from an OA

Utility function for inspecting SOAs obtainable from an OA

guide_SOAs_from_OA(s, nOA, mOA, tOA, el = tOA, ...)

Arguments

s

required; the unique number of levels of the columns of a given OA (need not be prime or prime power)

nOA

required; the number of runs of the OA

mOA

required; the number of columns of the OA

tOA

required; the strength of the OA; strengths larger than 5 are reduced to 5; el must not be larger than the (reduced) strength, except for tOA=2 with el=3, which is supported by the LLY algorithm

el

the power to which s is to be taken, i.e. the SOA will have columns with s^el levels; default: tOA.
except for tOA=2 and el=3, el can be chosen smaller than tOA, but not larger. If el is smaller than tOA, tOA is internally reduced before working out the possibilities.

...

currently unused

Value

The function returns a data frame, each row of which contains a possibility. There is example code for constructing the SOA. The code assumes that a given OA has the name OA; this can of course be modified by the user. Further code details can also be adjusted by the user (see the documentation of the respective functions).

Details

The function provides the possible creation variants of an SOA from a strength tOA OA with mOA s-level columns in nOA runs, for an SOA that has columns in s^el levels. Note that the SOA may have nOA runs or s*nOA runs, depending on the construction.

References

For full detail, see SOAs-package.

Groemping (2022)
He and Tang (2013)
He, Cheng and Tang (2018)
Liu and Liu (2015)
Li, Liu and Yang (2021)
Shi and Tang (2020)
Zhou and Tang (2019)

See also

Author

Ulrike Groemping

Examples

## guide_SOAs_from_OA
## there is an OA(81, 3^10, 3) (L81.3.10 in package DoE.base)
## inspect what can be done with it:
guide_SOAs_from_OA(s=3, mOA=10, nOA=81, tOA=3)
#>     type strength   n nlevels  m mmax orthogonal                  code
#> HT    HT        3  81      27  9    9         no         SOAs(OA, t=3)
#> LL    LL        3  81      27  4    4        yes OSOAs_LiuLiu(OA, t=3)
#> LLY  LLY        3 243      27 10   10        yes       OSOAs(OA, el=3)
## the output shows that a strength 3 OSOA
## with 4 columns of 27 levels each can be obtained in 81 runs
## and provides the necessary code (replace OA with L81.3.10)
##      optimize=FALSE reduces example run time
OSOAs_LiuLiu(L81.3.10, t=3, optimize=FALSE)
#>    [,1] [,2] [,3] [,4]
#> 1     0    2    0    2
#> 2     0    5   10   16
#> 3     0    8   20   21
#> 4     9    1   22    9
#> 5     9    4    5   26
#> 6     9    7   12    4
#> 7    18    0   17   19
#> 8    18    3   24    6
#> 9    18    6    7   14
#> 10    3    2    7   17
#> 11    3    5   17   22
#> 12    3    8   24    0
#> 13   12    1   20   24
#> 14   12    4    0    5
#> 15   12    7   10   10
#> 16   21    0   12    7
#> 17   21    3   22   12
#> 18   21    6    5   20
#> 19    6    2    5   23
#> 20    6    5   12    1
#> 21    6    8   22   15
#> 22   15    1   24    3
#> 23   15    4    7   11
#> 24   15    7   17   25
#> 25   24    0   10   13
#> 26   24    3   20   18
#> 27   24    6    0    8
#> 28    1   11   13   13
#> 29    1   14   23   18
#> 30    1   17    3    8
#> 31   10   10    8   23
#> 32   10   13   15    1
#> 33   10   16   25   15
#> 34   19    9   18    3
#> 35   19   12    1   11
#> 36   19   15   11   25
#> 37    4   11   11   19
#> 38    4   14   18    6
#> 39    4   17    1   14
#> 40   13   10    3    2
#> 41   13   13   13   16
#> 42   13   16   23   21
#> 43   22    9   25    9
#> 44   22   12    8   26
#> 45   22   15   15    4
#> 46    7   11   15    7
#> 47    7   14   25   12
#> 48    7   17    8   20
#> 49   16   10    1   17
#> 50   16   13   11   22
#> 51   16   16   18    0
#> 52   25    9   23   24
#> 53   25   12    3    5
#> 54   25   15   13   10
#> 55    2   20   26   24
#> 56    2   23    6    5
#> 57    2   26   16   10
#> 58   11   19    9    7
#> 59   11   22   19   12
#> 60   11   25    2   20
#> 61   20   18    4   17
#> 62   20   21   14   22
#> 63   20   24   21    0
#> 64    5   20   21    3
#> 65    5   23    4   11
#> 66    5   26   14   25
#> 67   14   19   16   13
#> 68   14   22   26   18
#> 69   14   25    6    8
#> 70   23   18    2   23
#> 71   23   21    9    1
#> 72   23   24   19   15
#> 73    8   20   19    9
#> 74    8   23    2   26
#> 75    8   26    9    4
#> 76   17   19   14   19
#> 77   17   22   21    6
#> 78   17   25    4   14
#> 79   26   18    6    2
#> 80   26   21   16   16
#> 81   26   24   26   21
#> OSOA, strength 3
## or that an SOA with 9 non-orthogonal columns can be obtained
## in the same number of runs
SOAs(L81.3.10, t=3)
#> Optimization 1 of 1 started
#> Optimization round 1 of 1 started
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9]
#>  [1,]   20   15    6    7   15   19   22    3   25
#>  [2,]   26   12    1   23   21   24    0   26    4
#>  [3,]   23    9    5    9    0   23   17   10   10
#>  [4,]   26   14   18    5   22    8   10    7   13
#>  [5,]   23   11   22   18    1    4   24   18   19
#>  [6,]   20   17   26   16   16    0    5   14    7
#>  [7,]   23   10   12    0    2   12    7    2    1
#>  [8,]   20   16   16   25   17   11   12   22   16
#>  [9,]   26   13   11   14   23   16   20   15   22
#> [10,]   22   18    3    2   20   13   24   19   10
#> [11,]   19   24    7   24    8    9    5   12   25
#> [12,]   25   21    2   13   14   17   10    8    4
#> [13,]   19   26   24    6    6   20   12   23    7
#> [14,]   25   23   19   22   12   25   20   16   13
#> [15,]   22   20   23   11   18   21    7    0   19
#> [16,]   25   22    9    4   13    6    0   24   22
#> [17,]   22   19   13   20   19    5   17   11    1
#> [18,]   19   25   17   15    7    1   22    4   16
#> [19,]   24    3    0    3    4    7   20   17    4
#> [20,]   21    0    4   19   10    3    7    1   10
#> [21,]   18    6    8   17   25    2   12   21   25
#> [22,]   21    2   21    1   11   14   17    9   19
#> [23,]   18    8   25   26   26   10   22    5    7
#> [24,]   24    5   20   12    5   15    0   25   13
#> [25,]   18    7   15    8   24   18    5   13   16
#> [26,]   24    4   10   21    3   26   10    6   22
#> [27,]   21    1   14   10    9   22   24   20    1
#> [28,]   14    9    3    2   19    3    8    9   18
#> [29,]   11   15    7   24    7    2   13    5    6
#> [30,]   17   12    2   13   13    7   18   25   12
#> [31,]   11   17   24    6    8   10   23   13   15
#> [32,]   17   14   19   22   14   15    1    6   21
#> [33,]   14   11   23   11   20   14   15   20    0
#> [34,]   17   13    9    4   12   26   11   17    3
#> [35,]   14   10   13   20   18   22   25    1    9
#> [36,]   11   16   17   15    6   18    3   21   24
#> [37,]   16   21    0    3    3   24    1    7   12
#> [38,]   13   18    4   19    9   23   15   18   18
#> [39,]   10   24    8   17   24   19   23   14    6
#> [40,]   13   20   21    1   10    4   25    2    0
#> [41,]   10   26   25   26   25    0    3   22   15
#> [42,]   16   23   20   12    4    8   11   15   21
#> [43,]   10   25   15    8   26   11   13    3   24
#> [44,]   16   22   10   21    5   16   18   26    3
#> [45,]   13   19   14   10   11   12    8   10    9
#> [46,]    9    6    6    7   17    9    3   23    6
#> [47,]   15    3    1   23   23   17   11   16   12
#> [48,]   12    0    5    9    2   13   25    0   18
#> [49,]   15    5   18    5   21   25   18   24   21
#> [50,]   12    2   22   18    0   21    8   11    0
#> [51,]    9    8   26   16   15   20   13    4   15
#> [52,]   12    1   12    0    1    5   15   19    9
#> [53,]    9    7   16   25   16    1   23   12   24
#> [54,]   15    4   11   14   22    6    1    8    3
#> [55,]    8   12    0    3    5   17    9   24   23
#> [56,]    5    9    4   19   11   13   26   11    2
#> [57,]    2   15    8   17   26    9    4    4   17
#> [58,]    5   11   21    1    9   21    6   19   11
#> [59,]    2   17   25   26   24   20   14   12   26
#> [60,]    8   14   20   12    3   25   19    8    5
#> [61,]    2   16   15    8   25    1   21   23    8
#> [62,]    8   13   10   21    4    6    2   16   14
#> [63,]    5   10   14   10   10    5   16    0   20
#> [64,]    1   24    6    7   16    2   14   13   17
#> [65,]    7   21    1   23   22    7   19    6   23
#> [66,]    4   18    5    9    1    3    6   20    2
#> [67,]    7   23   18    5   23   15    2   17    5
#> [68,]    4   20   22   18    2   14   16    1   11
#> [69,]    1   26   26   16   17   10   21   21   26
#> [70,]    4   19   12    0    0   22   26    9   20
#> [71,]    1   25   16   25   15   18    4    5    8
#> [72,]    7   22   11   14   21   26    9   25   14
#> [73,]    3    0    3    2   18   23   16    2    2
#> [74,]    0    6    7   24    6   19   21   22   17
#> [75,]    6    3    2   13   12   24    2   15   23
#> [76,]    0    8   24    6    7    0    4    3   26
#> [77,]    6    5   19   22   13    8    9   26    5
#> [78,]    3    2   23   11   19    4   26   10   11
#> [79,]    6    4    9    4   14   16   19    7   14
#> [80,]    3    1   13   20   20   12    6   18   20
#> [81,]    0    7   17   15    8   11   14   14    8
#> SOA, strength 3