functions to evaluate low order projection properties of (O)SOAs

soacheck2D and soacheck3D evaluate 2D and 3D projections, ocheck and ocheck3 evaluate pairwise or 3-orthogonality of columns, and count_npairs evaluates the number of level pairs in 2D projections

ocheck(D, verbose = FALSE)

ocheck3(D, verbose = FALSE)

count_npairs(D, minn = 1)

count_nallpairs(ns)

soacheck2D(D, s = 3, el = 3, t = 3, verbose = FALSE)

soacheck3D(D, s = 3, el = 3, t = 3, verbose = FALSE)

Arguments

D

a matrix with factor levels or an object of class SOA;
factor levels can start with 0 or with 1, and need to be consecutively numbered

verbose

logical; if TRUE, additional information is printed (confounded pair or triple projections with A2 or A3, respectively, or table of correlations)

minn

small integer number; the function counts pairs that are covered at least minn times

ns

vector of numbers of levels for each column

s

the prime or prime power according to which the array is checked

el

the exponent so that the number of levels of the array is s^el (if s is not NULL)

t

the strength for which to look (2, 3, or 4), equal to the sum of the exponents in the stratification dimensions; for example, soacheck2D considers
sxs 2D projections with t=2,
s^2xs and sxs^2 projections with t=3,
and s^3xs, s^2xs^2 and sxs^3 projections with t=4.
If t=4 and el=2, property gamma (s^3 x s and s x s^3) is obviously impossible and will not be part of the checks.

Value

Functions whose names contain "check" return a logical. Functions count_npairs and count_npairs return a vector of counts for level combinations in factor pairs (in the order of the columns of DoE.base:::nchoosek(ncol(D),2)), either for the array in D, or for designs with numbers of levels given in ns.

Details

Functions soacheck2D and soacheck3D inspect 2D and 3D stratification, respectively. Each column must have s^el levels. t specifies the degree of balance the functions are asked to look for.

Function soacheck2D,

with el=t=2, looks for strength 2 conditions (s^2 levels, sxs balance),

with el=2, t=3, looks for strength 2+ / 3- conditions (s^2 levels, s^2xs balance),

with el=t=3, looks for strength 2* / 3 conditions (s^3 levels, s^2xs balance).

with el=2, t=4, looks for the enhanced strength 2+ / 3- property alpha (s^2 levels, s^2xs^2 balance).

and with el=3, t=4, looks for strength 3+ / 4 conditions (s^3 levels, s^3xs and s^2xs^2 balance).

Function soacheck3D,

with el=2, t=3, looks for strength 3- conditions (s^2 levels, sxsxs balance),

with el=t=3, looks for strength 3 conditions (s^3 levels, sxsxs balance),

and with el=3, t=4, looks for strength 3+ / 4 conditions (s^3 levels, s^2xsxs balance).

If verbose=TRUE, the functions print the pairs or triples that violate the projection requirements for 2D or 3D.

References

For full detail, see SOAs-package.

Groemping (2021)
He and Tang (2013)
Shi and Tang (2020)

Author

Ulrike Groemping

Examples

nullcase <- matrix(0:7, nrow=8, ncol=4)
soacheck2D(nullcase, s=2)
#> [1] FALSE
soacheck3D(nullcase, s=2)
#> [1] FALSE

## Shi and Tang strength 3+ construction in 7 8-level factors for 32 runs
D <- SOAs_8level(32, optimize=FALSE)

## check for strength 3+ (default el=3 is OK)
## 2D check
soacheck2D(D, s=2, t=4)
#> [1] TRUE
## 3D check
soacheck3D(D, s=2, t=4)
#> [1] TRUE
## not an OSOA
ocheck(D)
#> [1] TRUE

## an OSOA of strength 3 with 3-orthogonality
## 4 columns in 27 levels each
## second order model matrix

D_o <- OSOAs_LiuLiu(DoE.base::L81.3.10, optimize=FALSE)
ocheck3(D_o)
#> [1] TRUE

## benefit of 3-orthogonality for second order linear models
colnames(D_o) <- paste0("X", 1:4)
y <- stats::rnorm(81)
mylm <- stats::lm(y~(X1+X2+X3+X4)^2 + I(X1^2)+I(X2^2)+I(X3^2)+I(X4^2),
                   data=as.data.frame(scale(D_o, scale=FALSE)))
crossprod(stats::model.matrix(mylm))
#>             (Intercept)   X1   X2   X3   X4 I(X1^2) I(X2^2) I(X3^2) I(X4^2)
#> (Intercept)          81    0    0    0    0    4914    4914    4914    4914
#> X1                    0 4914    0    0    0       0       0       0       0
#> X2                    0    0 4914    0    0       0       0       0       0
#> X3                    0    0    0 4914    0       0       0       0       0
#> X4                    0    0    0    0 4914       0       0       0       0
#> I(X1^2)            4914    0    0    0    0  535626  289368  315612  289368
#> I(X2^2)            4914    0    0    0    0  289368  535626  298116  289368
#> I(X3^2)            4914    0    0    0    0  315612  298116  535626  289368
#> I(X4^2)            4914    0    0    0    0  289368  289368  289368  535626
#> X1:X2                 0    0    0    0    0   38880  -38880       0   39852
#> X1:X3                 0    0    0    0    0    8748   74844    2916   -1458
#> X1:X4                 0    0    0    0    0       0   -4266   13014       0
#> X2:X3                 0    0    0    0    0   34668  -26244   26244    4212
#> X2:X4                 0    0    0    0    0    -486    2916   79218   -4374
#> X3:X4                 0    0    0    0    0   78246    -972   38880  -38880
#>              X1:X2  X1:X3  X1:X4  X2:X3  X2:X4  X3:X4
#> (Intercept)      0      0      0      0      0      0
#> X1               0      0      0      0      0      0
#> X2               0      0      0      0      0      0
#> X3               0      0      0      0      0      0
#> X4               0      0      0      0      0      0
#> I(X1^2)      38880   8748      0  34668   -486  78246
#> I(X2^2)     -38880  74844  -4266 -26244   2916   -972
#> I(X3^2)          0   2916  13014  26244  79218  38880
#> I(X4^2)      39852  -1458      0   4212  -4374 -38880
#> X1:X2       289368  34668   -486  74844  -4266   4401
#> X1:X3        34668 315612  78246      0   4401  13014
#> X1:X4         -486  78246 289368   4401  39852  -1458
#> X2:X3        74844      0   4401 298116   -972  79218
#> X2:X4        -4266   4401  39852   -972 289368   4212
#> X3:X4         4401  13014  -1458  79218   4212 289368