work horse function for SOAs

work horse function for SOAs

soa(oa, t = 3, m = NULL, permlist = NULL, random = TRUE)

Arguments

oa

matrix or data.frame that contains an ingoing symmetric OA. Levels must be denoted as 0 to s-1 or as 1 to s.

t

the strength the SOA should have, can be 2, 3, 4, or 5. Must not be larger than the strength of oa, but can be smaller. The resulting SOA will have s^t levels

m

the requested number of columns (see details for permitted numbers of columns)

permlist

list of lists of permutations

random

logical

Value

a matrix

Details

The function is the workhorse function for function SOAs.

Examples

oa <- DoE.base::L81.3.10 - 1
SOAs:::soa(oa, t=3)   ## 9 columns with 27 levels each
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9]
#>  [1,]    3   17    8   18   24    7    4   20   15
#>  [2,]    0   14    1    5   12    2   18   13   18
#>  [3,]    6   11    3   16    0    3   17    6    3
#>  [4,]    0   12   11   23   13    9   10   21    0
#>  [5,]    6    9   13    7    1   13    6   17   12
#>  [6,]    3   15   15    9   25   17   23    1   24
#>  [7,]    6   10   23   25    2   23   25   25   21
#>  [8,]    3   16   25    0   26   24   12    9    6
#>  [9,]    0   13   18   14   14   19    2    5    9
#> [10,]    8   20    5   26   11   22    6   15    3
#> [11,]    5   26    7    1    8   26   23    2   15
#> [12,]    2   23    0   12   23   18   10   22   18
#> [13,]    5   24   17   19    6    6   12   10   24
#> [14,]    2   21   10    3   21    1    2    3    0
#> [15,]    8   18   12   17    9    5   25   26   12
#> [16,]    2   22   20   21   22   11   18   14    9
#> [17,]    8   19   22    8   10   12   17    7   21
#> [18,]    5   25   24   10    7   16    4   18    6
#> [19,]    1    5    2   22    4   10    2    4   18
#> [20,]    7    2    4    6   19   14   25   24    3
#> [21,]    4    8    6   11   16   15   12   11   15
#> [22,]    7    0   14   24   20   21   17    8   12
#> [23,]    4    6   16    2   17   25    4   19   24
#> [24,]    1    3    9   13    5   20   18   12    0
#> [25,]    4    7   26   20   15    8   23    0    6
#> [26,]    1    4   19    4    3    0   10   23    9
#> [27,]    7    1   21   15   18    4    6   16   21
#> [28,]   24   11    5   26   10   14   26    8   14
#> [29,]   21   17    7    1    7   15   13   19   26
#> [30,]   18   14    0   12   22   10    0   12    2
#> [31,]   21   15   17   19    8   25    5    0    8
#> [32,]   18   12   10    3   23   20   19   23   11
#> [33,]   24    9   12   17   11   21   15   16   23
#> [34,]   18   13   20   21   21    0   11    4   20
#> [35,]   24   10   22    8    9    4    7   24    5
#> [36,]   21   16   24   10    6    8   21   11   17
#> [37,]   20   23    2   22    3    2   19   21    2
#> [38,]   26   20    4    6   18    3   15   17   14
#> [39,]   23   26    6   11   15    7    5    1   26
#> [40,]   26   18   14   24   19   13    7   25   23
#> [41,]   23   24   16    2   16   17   21    9    8
#> [42,]   20   21    9   13    4    9   11    5   11
#> [43,]   23   25   26   20   17   24   13   20   17
#> [44,]   20   22   19    4    5   19    0   13   20
#> [45,]   26   19   21   15   20   23   26    6    5
#> [46,]   22    8    8   18   26   26   21   10   26
#> [47,]   19    5    1    5   14   18   11    3    2
#> [48,]   25    2    3   16    2   22    7   26   14
#> [49,]   19    3   11   23   12    1    0   14   11
#> [50,]   25    0   13    7    0    5   26    7   23
#> [51,]   22    6   15    9   24    6   13   18    8
#> [52,]   25    1   23   25    1   12   15   15    5
#> [53,]   22    7   25    0   25   16    5    2   17
#> [54,]   19    4   18   14   13   11   19   22   20
#> [55,]    9   14    2   22    5   18    9   14   10
#> [56,]   15   11    4    6   20   22    8    7   22
#> [57,]   12   17    6   11   17   26   22   18    7
#> [58,]   15    9   14   24   18    5   24   15    4
#> [59,]   12   15   16    2   15    6   14    2   16
#> [60,]    9   12    9   13    3    1    1   22   19
#> [61,]   12   16   26   20   16   16    3   10   25
#> [62,]    9   13   19    4    4   11   20    3    1
#> [63,]   15   10   21   15   19   12   16   26   13
#> [64,]   14   26    8   18   25   15   14    0    7
#> [65,]   11   23    1    5   13   10    1   23   10
#> [66,]   17   20    3   16    1   14   24   16   22
#> [67,]   11   21   11   23   14   20   20    4   19
#> [68,]   17   18   13    7    2   21   16   24    4
#> [69,]   14   24   15    9   26   25    3   11   16
#> [70,]   17   19   23   25    0    4    8    8   13
#> [71,]   14   25   25    0   24    8   22   19   25
#> [72,]   11   22   18   14   12    0    9   12    1
#> [73,]   16    2    5   26    9    3   16   25   22
#> [74,]   13    8    7    1    6    7    3    9    7
#> [75,]   10    5    0   12   21    2   20    5   10
#> [76,]   13    6   17   19    7   17   22   20   16
#> [77,]   10    3   10    3   22    9    9   13   19
#> [78,]   16    0   12   17   10   13    8    6    4
#> [79,]   10    4   20   21   23   19    1   21    1
#> [80,]   16    1   22    8   11   23   24   17   13
#> [81,]   13    7   24   10    8   24   14    1   25