work horse function for SOAs

work horse function for SOAs

soa(oa, t = 3, m = NULL, permlist = NULL, random = TRUE)

Arguments

oa
matrix or data.frame that contains an ingoing symmetric OA. Levels must be denoted as 0 to s-1 or as 1 to s.

t
the strength the SOA should have, can be 2, 3, 4, or 5. Must not be larger than the strength of oa, but can be smaller. The resulting SOA will have s^t levels

m
the requested number of columns (see details for permitted numbers of columns)

permlist
list of lists of permutations

random
logical

Value

a matrix

Details

The function is the workhorse function for function SOAs.

Examples

oa <- DoE.base::L81.3.10 - 1
SOAs:::soa(oa, t=3)   ## 9 columns with 27 levels each

#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9]
#>  [1,]   22   18   17   22    9   18   26    5   18
#>  [2,]   25   21   13   15    6   22    9   10    6
#>  [3,]   19   24    9    2   21   26    4   24   12
#>  [4,]   25   22   23   24    7    5    2    0   15
#>  [5,]   19   25   19   11   22    6   21   17   21
#>  [6,]   22   19   24    4   10    1   16   22    0
#>  [7,]   19   26    2   20   23   16   14    7    3
#>  [8,]   22   20    7   13   11   11    6   12    9
#>  [9,]   25   23    3    6    8   12   19   20   24
#> [10,]   18    6   11   18    5   15   21   15   12
#> [11,]   21    0   16   14   20   10   16   23   18
#> [12,]   24    3   12    7   17   14    2    1    6
#> [13,]   21    1   26   23   18   20    6   13    0
#> [14,]   24    4   22   16   15   21   19   18   15
#> [15,]   18    7   18    0    3   25   14    8   21
#> [16,]   24    5    5   25   16    4    9   11   24
#> [17,]   18    8    1    9    4    8    4   25    3
#> [18,]   21    2    6    5   19    0   26    3    9
#> [19,]   26   12   14   26   25    3   19   19    6
#> [20,]   20   15   10   10   13    7   14    6   12
#> [21,]   23    9   15    3    1    2    6   14   18
#> [22,]   20   16   20   19   14   17    4   26   21
#> [23,]   23   10   25   12    2    9   26    4    0
#> [24,]   26   13   21    8   26   13    9    9   15
#> [25,]   23   11    8   21    0   19   16   21    9
#> [26,]   26   14    4   17   24   23    2    2   24
#> [27,]   20   17    0    1   12   24   21   16    3
#> [28,]    1   24   11   18    4    7   13   26   22
#> [29,]    4   18   16   14   19    2    8    4    1
#> [30,]    7   21   12    7   16    3   18    9   16
#> [31,]    4   19   26   23   20    9   25   21   10
#> [32,]    7   22   22   16   17   13   11    2   25
#> [33,]    1   25   18    0    5   17    3   16    4
#> [34,]    7   23    5   25   15   23    1   19    7
#> [35,]    1   26    1    9    3   24   23    6   13
#> [36,]    4   20    6    5   18   19   15   14   19
#> [37,]    6    3   14   26   24   22   11    0   16
#> [38,]    0    6   10   10   12   26    3   17   22
#> [39,]    3    0   15    3    0   18   25   22    1
#> [40,]    0    7   20   19   13    6   23    7    4
#> [41,]    3    1   25   12    1    1   15   12   10
#> [42,]    6    4   21    8   25    5    1   20   25
#> [43,]    3    2    8   21    2   11    8    5   19
#> [44,]    6    5    4   17   26   12   18   10    7
#> [45,]    0    8    0    1   14   16   13   24   13
#> [46,]    5    9   17   22   11   10   15   13    1
#> [47,]    8   12   13   15    8   14    1   18   16
#> [48,]    2   15    9    2   23   15   23    8   22
#> [49,]    8   13   23   24    6   21   18   11   25
#> [50,]    2   16   19   11   21   25   13   25    4
#> [51,]    5   10   24    4    9   20    8    3   10
#> [52,]    2   17    2   20   22    8    3   15   13
#> [53,]    5   11    7   13   10    0   25   23   19
#> [54,]    8   14    3    6    7    4   11    1    7
#> [55,]   16   21   14   26   26   14    0   11   26
#> [56,]   10   24   10   10   14   15   22   25    5
#> [57,]   13   18   15    3    2   10   17    3   11
#> [58,]   10   25   20   19   12   25   12   15   14
#> [59,]   13   19   25   12    0   20    7   23   20
#> [60,]   16   22   21    8   24   21   20    1    8
#> [61,]   13   20    8   21    1    0   24   13    2
#> [62,]   16   23    4   17   25    4   10   18   17
#> [63,]   10   26    0    1   13    8    5    8   23
#> [64,]   12    0   17   22   10    2    7   21   11
#> [65,]   15    3   13   15    7    3   20    2   26
#> [66,]    9    6    9    2   22    7   12   16    5
#> [67,]   15    4   23   24    8   13   10   19    8
#> [68,]    9    7   19   11   23   17    5    6   14
#> [69,]   12    1   24    4   11    9   24   14   20
#> [70,]    9    8    2   20   21   24   22   26   23
#> [71,]   12    2    7   13    9   19   17    4    2
#> [72,]   15    5    3    6    6   23    0    9   17
#> [73,]   11   15   11   18    3   26    5    7    5
#> [74,]   14    9   16   14   18   18   24   12   11
#> [75,]   17   12   12    7   15   22   10   20   26
#> [76,]   14   10   26   23   19    1   17    5   20
#> [77,]   17   13   22   16   16    5    0   10    8
#> [78,]   11   16   18    0    4    6   22   24   14
#> [79,]   17   14    5   25   17   12   20    0   17
#> [80,]   11   17    1    9    5   16   12   17   23
#> [81,]   14   11    6    5   20   11    7   22    2